% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClustMIC.R
\name{ClustMIC}
\alias{ClustMIC}
\title{Unsupervised clustering on (GPL) intensities and associated MIC indexes}
\usage{
ClustMIC(Intensities, nClust, Trcl, Dendr = TRUE)
}
\arguments{
\item{Intensities}{A numeric matrix of intensities (can be a pTreatGPL object).}

\item{nClust}{The number of groups to retrieve (donors, mixtures, ...).}

\item{Trcl}{The real groups' memberships of the samples, true class.}

\item{Dendr}{Logical argument (TRUE/FALSE) to obtain graphical dendrogram based on the Ward algorithm.}
}
\value{
A list of MIC quality indexes (Dunn, Davies-Bouldin, Rand and adjusted-Rand):
\describe{
  \item{\code{DunnW}}{Dunn index for Ward clustering}
  \item{\code{DunnKM}}{Dunn index for Kmeans clustering}
  \item{\code{DBW}}{Davies-Bouldin index for Ward clustering}
  \item{\code{DBKM}}{Davies-Bouldin index for Kmeans clustering}
  \item{\code{RandW}}{Rand index for Ward clustering}
  \item{\code{RandKM}}{Rand index for Kmeans clustering}
  \item{\code{AdjRandW}}{Adjusted Rand index for Ward clustering}
  \item{\code{AdjRandKM}}{Adjusted Rand index for Kmeans clustering}
   }
}
\description{
Unsupervised clustering on (GPL) intensities based on Ward and K-Means algorithms.
Calculation of MIC statistical criteria of clustering quality:
Dunn, Davies-Bouldin, Rand and adjusted-Rand indexes.
}
\examples{
data('HumanSerum')
ClustMIC(Intensities = HumanSerumSpectra, nClust = 4, Trcl = ClassHS, Dendr = TRUE)

}
\author{
Baptiste Feraud, Manon Martin
}
