% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OPLDA.R
\name{cvOPLSDA}
\alias{cvOPLSDA}
\title{K-fold cross-validation for OPLS-DA.}
\usage{
cvOPLSDA(x, y, k_fold = 10, NumOrtho = 1)
}
\arguments{
\item{x}{A data matrix on which will be based the analysis.}

\item{y}{A numerical vector representing the class of individuals.}

\item{k_fold}{The number of sub-datasets to create for cross-validation.}

\item{NumOrtho}{The maximum number of orthogonal components allowed in OPLSDA.}
}
\value{
A list with the following elements:
\describe{
  \item{\code{RMSECV}}{Result from OPLS-DA analysis with the OPLSDA function.}
  \item{\code{folds_i}}{A vector indicating the group of observations for cross-validation.}
  \item{\code{Prop1}}{Table of proportions of 1's in y for each subgroup.}
}
}
\description{
K-fold cross-validation for OPLS-DA based on the RMSE criterion and stratified according to y.
}
\examples{
data("DataSimul")
x = DataSimul[["x"]]
y = DataSimul[["y"]]
cvOPLSDA.res = cvOPLSDA(x = x, y = y, k_fold = 10, NumOrtho = 5)
plot(cvOPLSDA.res$RMSECV)

}

