% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DrawPCA.R
\name{DrawPCA}
\alias{DrawPCA}
\title{Scores and Loadings plots}
\usage{
DrawPCA(PCAobj, drawNames = TRUE, createWindow = FALSE, main = NULL,
  class = NULL, axes = c(1, 2), type = c("scores", "loadings"),
  loadingstype = c("l", "p"), num.stacked = 4, xlab = NULL)
}
\arguments{
\item{PCAobj}{The objects resulting from a PCA analysis with the SVDforPCA function.}

\item{drawNames}{If \code{TRUE}, will show the observations names on the Scores plot.}

\item{createWindow}{If \code{TRUE}, will create a new window for the plot.}

\item{main}{Plot title. If \code{NULL}, default title is provided.}

\item{class}{Optional numeric vector giving the class of the observations.}

\item{axes}{Numerical vector indicating the PC axes that are drawn. Only the two first values are considered for Scores plot and only the first is considered for Loadings plot.}

\item{type}{The type of plot, either \code{"scores"} or \code{"loadings"}}

\item{loadingstype}{The type of Loadings plot, either a line plot (\code{"l"}) or points with histogram-like vertical lines (\code{"p"}).}

\item{num.stacked}{Number of stacked plots if \code{type} is \code{"loadings"}.}

\item{xlab}{Label of the x-axis.}
}
\value{
A score or loading plot in the current device.
}
\description{
Draws scores and Loadings plots for PCA analysis derived from the SVDforPCA function.
}
\examples{

data("HumanSerum")
PCAobj = SVDforPCA(HumanSerumSpectra)
class = ClassHS

DrawPCA(PCAobj, drawNames=TRUE,
createWindow=FALSE, main = "PCA score plot for HumanSerum dataset",
  class = class, axes =c(1,2), type ="scores")

DrawPCA(PCAobj, drawNames=TRUE,
createWindow=FALSE, main = "PCA loadings plot for HumanSerum dataset",
   axes = 1, type ="loadings", loadingstype="l")

}

