% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DrawSL.R
\name{DrawSL}
\alias{DrawSL}
\title{Scores and Loadings plots}
\usage{
DrawSL(obj, type.obj = c("PCA", "PLSDA", "OPLSDA"), drawNames = TRUE,
  createWindow = FALSE, main = NULL, class = NULL, axes = c(1, 2),
  type.graph = c("scores", "loadings"), loadingstype = c("l", "p"),
  num.stacked = 4, xlab = NULL, ang = 0)
}
\arguments{
\item{obj}{The objects resulting from a PCA analysis with the SVDforPCA function.}

\item{type.obj}{The type of object to be plotted.}

\item{drawNames}{If \code{TRUE}, will show the observations names on the Scores plot.}

\item{createWindow}{If \code{TRUE}, will create a new window for the plot.}

\item{main}{Plot title. If \code{NULL}, default title is provided.}

\item{class}{Optional numeric vector giving the class of the observations.}

\item{axes}{Numerical vector indicating the PC axes that are drawn. Only the two first values are considered for scores plot.}

\item{type.graph}{The type of plot, either \code{"scores"} or \code{"loadings"}}

\item{loadingstype}{The type of Loadings plot, either a line plot (\code{"l"}) or points with histogram-like vertical lines (\code{"p"}).}

\item{num.stacked}{Number of stacked plots if \code{type} is \code{"loadings"}.}

\item{xlab}{Label of the x-axis.}

\item{ang}{Angle to rotate the x axis labels for a better visualisation.}
}
\value{
A score or loading plot in the current device.
}
\description{
Draws scores and Loadings plots for PCA analysis derived from the SVDforPCA function.
}
\examples{

data("HumanSerum")
res.PCA = SVDforPCA(HumanSerumSpectra)
class = ClassHS

DrawSL(res.PCA, drawNames=TRUE, type.obj = "PCA",
createWindow=FALSE, main = "PCA score plot for HumanSerum dataset",
  class = class, axes =c(1,2), type.graph ="scores")

DrawSL(res.PCA, drawNames=TRUE, type.obj = "PCA",
createWindow=FALSE, main = "PCA loadings plot for HumanSerum dataset",
   axes = 1, type.graph ="loadings", loadingstype="l")

}

