% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index.DB.R
\name{index.DB}
\alias{index.DB}
\title{Function from ClusterSim (version 0.49-2)}
\usage{
index.DB(x, cl, d = NULL, centrotypes = "centroids", p = 2, q = 2)
}
\arguments{
\item{x}{data}

\item{cl}{vector of integers indicating the cluster to which each object is allocated}

\item{d}{optional distance matrix, used for calculations if centrotypes="medoids"}

\item{centrotypes}{"centroids" or "medoids"}

\item{p}{the power of the Minkowski distance between centroids or medoids of clusters: p=1 - Manhattan distance; p=2 - Euclidean distance}

\item{q}{the power of dispersion measure of a cluster: q=1 - the average distance of objects in the r-th cluster to the centroid or medoid of the r-th cluster; q=2 - the standard deviation of the distance of objects in the r-th cluster to the centroid or medoid of the r-th cluster}
}
\description{
Function from ClusterSim (version 0.49-2)
}
