% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Inertia.R
\name{Inertia}
\alias{Inertia}
\title{Measure of inertia}
\usage{
Inertia(x, y, print = FALSE)
}
\arguments{
\item{x}{A numeric matrix.}

\item{y}{A numeric vector of data clusters.}

\item{print}{If \code{TRUE}, prints the inertia values}
}
\value{
A list with the following elements:
\describe{
  \item{\code{Name.group}}{Group membership}
  \item{\code{InertiaT}}{Total inertia}
  \item{\code{InertiaB}}{Between inertia}
  \item{\code{InertiaW}}{Within inertia}
}
}
\description{
Calculates the within, between and total inertia of a data matrix
}
\examples{
data("iris")
data = as.matrix((iris[,1:4]))
class = as.numeric(iris[,5])

Res = Inertia(x = data, y = class, print = TRUE)

}

