% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LinePlot.R
\name{LinePlot}
\alias{LinePlot}
\title{Line plots}
\usage{
LinePlot(
  X,
  createWindow = FALSE,
  main = NULL,
  rows = NULL,
  type = c("l", "p", "s"),
  num.stacked = 4,
  xlab = NULL,
  ylab = NULL,
  ang = c("0", "45", "90"),
  xaxis_type = c("numerical", "character"),
  nxaxis = 10,
  y_axis_text = TRUE,
  xaxis_size = 11
)
}
\arguments{
\item{X}{A numerical matrix containing the rows to be drawn.}

\item{createWindow}{If \code{TRUE}, will create a new window for the plot.}

\item{main}{Plot title. If \code{NULL}, default title is provided.}

\item{rows}{Numerical vector indicating the X matrix rows that are drawn.}

\item{type}{The type of plot, either a line plot (\code{'l'}), points (\code{'p'}) or segments (\code{'s'}).}

\item{num.stacked}{Number of stacked plots.}

\item{xlab}{Label for the x-axis.}

\item{ylab}{Label for the y-axis.}

\item{ang}{Angle to rotate the x axis labels for a better visualisation, either 0, 45 or 90 degrees.}

\item{xaxis_type}{Specify if the xaxis is numerical or character (corresponds to the colnames of X).}

\item{nxaxis}{Number of thick marks on the xaxis for a character x variable.}

\item{y_axis_text}{if \code{TRUE}, will display the y-axis text.}

\item{xaxis_size}{Size of the x axis.}
}
\value{
A loading plot in the current device.
}
\description{
Draws Line plots.
}
\details{
Better results printping is obtained by saving the plot as an object
}
\examples{

X <- rbind(sample(1:20,20), sample(21:40,20), sample(41:60,20))
colnames(X) <- as.character(1:20)

a <- LinePlot(X = X, createWindow = FALSE, main = "line plot",  rows = c(1, 2),
             type = "l", num.stacked = 4, xlab = "x-axis", ylab = "y-axis",
             ang = "0", xaxis_type = "numerical", nxaxis = 10)

}
