\name{MBXUCL-internal}
\alias{A2R}

\title{Internal MBXUCL Functions}
\description{
  Internal MBXUCL Functions
}
\details{
  These are not to be called by the user (or in some cases are just
                                          waiting for proper documentation to be written ...).
}
\keyword{internal}
