% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OPLDA.R
\name{OPLSDA_pred}
\alias{OPLSDA_pred}
\title{Prediction for OPLS-DA}
\usage{
OPLSDA_pred(ropls, x.new)
}
\arguments{
\item{ropls}{Result from OPLS-DA analysis with the OPLSDA function.}

\item{x.new}{A vector or matrix of new observations.}
}
\value{
A vector with predicted y values.
}
\description{
Prediction of a new set of observations based on a built model.
}
\examples{
data("DataSimul")
x = DataSimul[["x"]]
y = DataSimul[["y"]]
oplsda.res = OPLSDA(x=x[-c(1:5),], y=y[-c(1:5)],
         impT = FALSE,impG = FALSE, no=2, nb = 15, out.path = ".")
OPLSDA_pred(ropls = oplsda.res, x.new = x[c(1:5),])
}

