% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PLSDA.R
\name{PLSDA}
\alias{PLSDA}
\title{PLS-DA}
\usage{
PLSDA(x, y, nLV = NULL, drawRMSEP = TRUE)
}
\arguments{
\item{x}{A numeric matrix of spectral intensities.}

\item{y}{The real groups' memberships of the samples.}

\item{nLV}{The number of latent variables for PLS-DA.}

\item{drawRMSEP}{Logical, if \code{"TRUE"}, will draw plot of RMSEP values to decide on the number of latent variables.}
}
\value{
A list with the following elements:
\describe{
  \item{\code{original.dataset}}{Original matrix of spectral intensities}
  \item{\code{Y}}{Dummy matrix representing the classes}
  \item{\code{nLV}}{Number of latent variables kept. If \code{NULL}, will be set to 3 in the finla model.}
  \item{\code{RMSEP}}{Root mean squared error of prediction }
  \item{\code{R2}}{Coefficient of multiple determination}
  \item{\code{Q2}}{Cross-validated coefficient of multiple determination}
  \item{\code{Q2cum}}{Cumulative cross-validated coefficient of multiple determination (over all response variables)}
  \item{\code{ExpVarY}}{Proportion of Y variance explained by the model}
  \item{\code{ExpVarX}}{Proportion of X variance explained by the model}
  \item{\code{coefficients}}{PLSDA coefficients}
  \item{\code{scores}}{PLSDA scores}
  \item{\code{loadings}}{PLSDA loadings}
   }
}
\description{
PLS-DA model for multiple classes (2 or more).
}
\examples{
data("HumanSerum")
PLSDA = function(x = HumanSerumSpectra, y = ClassHS, nLV = NULL, drawRMSEP = TRUE)

}
\author{
Manon Martin
}

