% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pTreatGPL.R
\name{pTreatGPL}
\alias{pTreatGPL}
\title{2D-NMR Global Peak List pre-treatments}
\usage{
pTreatGPL(Intensities, Coordinates, LowW, UpW)
}
\arguments{
\item{Intensities}{A numeric matrix of intensities (from individual GPL object).}

\item{Coordinates}{A numeric matrix of coordinates (C1, C2).}

\item{LowW}{Number corresponding to the lower value of the water zone to suppress (4.63).}

\item{UpW}{Number corresponding to the upper value of the water zone to suppress (5.1).}
}
\value{
A pre-treated matrix of intensities (with rows = spectra)
}
\description{
Elimination of negative intensities (=0), Elimination of the water zone,
Normalization of the intensities (sum = 1)
}
\examples{
path <-  system.file('extdata', package = 'MBXUCL')
GPL = read.table(file.path(path, 'UrineGPL.csv'), sep=',',
 header=TRUE, row.names = 1, na.strings = '.' )

Int = GPL[,seq(3,dim(GPL)[2]- 1, 2)]
Coord = GPL[,c(1,2)]
IntT <- pTreatGPL(Int, Coord, 4.63, 5.1)
IntT[,1:2]

}
\author{
Baptiste Feraud
}
