% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UnsupClustering.R
\name{UnsupClustering}
\alias{UnsupClustering}
\title{Unsupervised classification of spectra}
\usage{
UnsupClustering(x, RowNames = NULL, nClust = 3, y, print = FALSE,
  impG = FALSE, out.path = ".")
}
\arguments{
\item{x}{Datamatrix of spectra}

\item{RowNames}{Optional, give the rownames of x}

\item{nClust}{Number of clusters}

\item{y}{The true class of spectra (as numeric)}

\item{print}{If \code{TRUE}, prints the clustering results}

\item{impG}{If \code{TRUE}, will save graphics (pdf format)}

\item{out.path}{Path to save the graphs}
}
\value{
A list with the following elements:
\describe{
  \item{\code{m}}{ Number of columns of x}
  \item{\code{DI.Ward}}{Dunn index for Ward clustering}
  \item{\code{DI.Kmeans}}{Dunn index for Kmeans clustering}
  \item{\code{DBI.Wald}}{Davies-Bouldin index for Ward clustering}
  \item{\code{DBI.Kmeans}}{Davies-Bouldin index for Kmeans clustering}
  \item{\code{RI.Ward}}{Rand index for Ward clustering}
  \item{\code{RI.Kmeans}}{Rand index for Kmeans clustering}
  \item{\code{ARI.Ward}}{Adjusted Rand index for Ward clustering}
  \item{\code{ARI.Kmeans}}{Adjusted Rand index for Kmeans clustering}
   }
}
\description{
Runs Ward and K-means clustering algorithms and computes quality indexes from these clustering
}
\examples{
data("iris")
data = as.matrix((iris[,1:4]))
class = as.numeric(iris[,5])

Res = UnsupClustering(x = data, nClust = 3, y = class)


}

