% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DrawLoadings.R
\name{DrawLoadings}
\alias{DrawLoadings}
\title{Loadings plots}
\usage{
DrawLoadings(obj, type.obj = c("PCA", "PLSDA", "OPLSDA"),
  createWindow = FALSE, main = NULL, axes = c(1, 2),
  loadingstype = c("l", "p", "s"), num.stacked = 4, xlab = NULL,
  ylab = NULL, ang = c("0", "45", "90"), xaxis_type = c("numerical",
  "character"), nxaxis = 10, hline = NULL, xaxis_size = 11)
}
\arguments{
\item{obj}{The objects resulting from a PCA, PLSDA or OPLSDA analysis.}

\item{type.obj}{The type of object to be plotted.}

\item{createWindow}{If \code{TRUE}, will create a new window for the plot.}

\item{main}{Plot title. If \code{NULL}, default title is provided.}

\item{axes}{Numerical vector indicating the PC axes that are drawn.}

\item{loadingstype}{The type of Loadings plot, either a line plot (\code{'l'}), points (\code{'p'}) or segments (\code{'s'}).}

\item{num.stacked}{Number of stacked plots.}

\item{xlab}{Label for the x-axis.}

\item{ylab}{Label for the y-axis.}

\item{ang}{Angle to rotate the x axis labels for a better visualisation, either 0, 45 or 90 degrees.}

\item{xaxis_type}{Specify if the xaxis is numerical or character}

\item{nxaxis}{Number of thick marks on the xaxis for a character x variable}

\item{hline}{Numerical scalar. If not \code{NULL}, an horizontal dashed line is drawn at the \code{hline} value}

\item{xaxis_size}{Size of the x axis.}
}
\value{
A loading plot in the current device.
}
\description{
Draws Loadings plots for the SVDforPCA, PLSDA or OPLSDA functions.
}
\details{
Better results printping is obtained by saving the plot as an object
}
\examples{

data('HumanSerum')
res.PCA = SVDforPCA(HumanSerumSpectra)

DrawLoadings(res.PCA, type.obj = 'PCA',
createWindow = FALSE, main = 'PCA loadings plot for HumanSerum dataset',
   axes = 1,  loadingstype='l', xlab = "L1")

}
