% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DrawScores.R
\name{DrawScores}
\alias{DrawScores}
\title{Scores plots}
\usage{
DrawScores(obj, type.obj = c("PCA", "PLSDA", "OPLSDA"),
  drawNames = TRUE, createWindow = FALSE, main = NULL,
  color = NULL, pch = NULL, size = 1, cex.lab = 3, axes = c(1,
  2), xlab = NULL, ylab = NULL, drawEllipses = FALSE,
  typeEl = "norm", levelEl = 0.9, drawPolygon = FALSE,
  noLegend = FALSE, legend_color_manual = NULL,
  legend_shape_manual = NULL)
}
\arguments{
\item{obj}{The objects resulting from a PCA, PLSDA or OPLSDA analysis.}

\item{type.obj}{The type of object to be plotted.}

\item{drawNames}{If \code{TRUE}, will show the observations names on the scores plot.}

\item{createWindow}{If \code{TRUE}, will create a new window for the plot.}

\item{main}{Plot title. If \code{NULL}, default title is provided.}

\item{color}{Optional character, factor or numeric vector giving the color of the observations. If \code{length(color)} = 1, the unique color is kept for all the points.}

\item{pch}{Optional character, factor or numeric vector giving the pch of the observations.}

\item{size}{The points size.}

\item{cex.lab}{The size of points labels.}

\item{axes}{Numerical vector indicating the PC axes that are drawn. Only the two first values are considered for scores plot. See details#' @param num.stacked Number of stacked plots if \code{type} is \code{'loadings'}.}

\item{xlab}{Label for the x-axis.}

\item{ylab}{Label for the y-axis.}

\item{drawEllipses}{If \code{TRUE}, will draw ellipses with the \code{ggplot2::stat_ellipse} with groups coresponding to the color vector.}

\item{typeEl}{The type of ellipse, either "norm" (multivariate normal distribution), "t" (multivariate t-distribution) and "euclid" draws a circle with the radius equal to level, representing the euclidean distance from the center.}

\item{levelEl}{The confidence level at which to draw an ellipse.}

\item{drawPolygon}{If \code{TRUE}, will relate the points linked to the same group (color vector)}

\item{noLegend}{If \code{TRUE}, no legend is drawn.}

\item{legend_color_manual}{If not \code{NULL}, a named character vector giving manually the colors to be drawn, named with the levels of the color vector.}

\item{legend_shape_manual}{If not \code{NULL}, a named character vector giving manually the shapes to be drawn, named with the levels of the pch vector.}
}
\value{
A score or loading plot in the current device.
}
\description{
Draws scores plots for the SVDforPCA, PLSDA or OPLSDA functions.
}
\details{
If \code{type.obj} is \code{'OPLSDA'}, axes = 1 represents the predictive score vector, axes = 2 represents the first orthogonal score vector, etc.
}
\examples{

data('HumanSerum')
res.PCA = SVDforPCA(HumanSerumSpectra)
class = ClassHS

DrawScores(res.PCA, drawNames=TRUE, type.obj = 'PCA',
createWindow=FALSE, main = 'PCA score plot for HumanSerum dataset',
  color = class, axes =c(1,2))


}
