% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DrawSL.R
\name{DrawSL}
\alias{DrawSL}
\title{Scores and Loadings plots}
\usage{
DrawSL(obj, type.obj = c("PCA", "PLSDA", "OPLSDA"), drawNames = TRUE,
  createWindow = FALSE, main = NULL, class = NULL, axes = c(1, 2),
  type.graph = c("scores", "loadings"), loadingstype = c("l", "p",
  "s"), num.stacked = 4, xlab = NULL, ylab = NULL, ang = 0,
  xaxis = c("numerical", "character"), nxaxis = 10)
}
\arguments{
\item{obj}{The objects resulting from a PCA analysis with the SVDforPCA function.}

\item{type.obj}{The type of object to be plotted.}

\item{drawNames}{If \code{TRUE}, will show the observations names on the Scores plot.}

\item{createWindow}{If \code{TRUE}, will create a new window for the plot.}

\item{main}{Plot title. If \code{NULL}, default title is provided.}

\item{class}{Optional character or numeric vector giving the class of the observations.}

\item{axes}{Numerical vector indicating the PC axes that are drawn. Only the two first values are considered for scores plot. See details}

\item{type.graph}{The type of plot, either \code{'scores'} or \code{'loadings'}}

\item{loadingstype}{The type of Loadings plot, either a line plot (\code{'l'}), points (\code{'p'}) or segments (\code{'s'}).}

\item{num.stacked}{Number of stacked plots if \code{type} is \code{'loadings'}.}

\item{xlab}{Label for the x-axis.}

\item{ylab}{Label for the y-axis.}

\item{ang}{Angle to rotate the x axis labels for a better visualisation.}

\item{xaxis}{Specify if the xaxis is numerical or character}

\item{nxaxis}{Number of thick marks on the xaxis for a character x variable}
}
\value{
A score or loading plot in the current device.
}
\description{
Draws scores and Loadings plots for PCA analysis derived from the SVDforPCA function.
}
\details{
If \code{type.obj} is \code{'OPLSDA'}, axes = 1 represents the predictive score vector, axes = 2 represents the first orthogonal score vector, etc.
}
\examples{

data('HumanSerum')
res.PCA = SVDforPCA(HumanSerumSpectra)
class = ClassHS

DrawSL(res.PCA, drawNames=TRUE, type.obj = 'PCA',
createWindow=FALSE, main = 'PCA score plot for HumanSerum dataset',
  class = class, axes =c(1,2), type.graph ='scores')

DrawSL(res.PCA, drawNames=TRUE, type.obj = 'PCA',
createWindow=FALSE, main = 'PCA loadings plot for HumanSerum dataset',
   axes = 1, type.graph ='loadings', loadingstype='l')

}
