% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Inertia.R
\name{Inertia}
\alias{Inertia}
\title{Measure of inertia}
\usage{
Inertia(x, y, print = FALSE)
}
\arguments{
\item{x}{A numeric matrix.}

\item{y}{A numeric vector of data clusters.}

\item{print}{If \code{TRUE}, prints the inertia values}
}
\value{
A list with the following elements:
\describe{
  \item{\code{Between_within}}{A matrix with the following elements: Name.group = Group membership,
  InertiaT = Total inertia, InertiaB = Between inertia, InertiaW = Within inertia.}
  \item{\code{Per_group}}{A matrix with the following elements: Inertia_group = Inertia per group,
  Inertia_group100 =  Inertia per group (percentage of ITOT pergroup), Inertia_moy_group = Mean Inertia per group.}
}
}
\description{
Calculates the within, between and total inertia of a data matrix
}
\examples{
data("HumanSerum")
Res = Inertia(x = HumanSerumSpectra, y = ClassHS, print = TRUE)

}

