% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LsOPLS.R
\name{LsOPLS}
\alias{LsOPLS}
\title{Light-sparse-OPLS method for biomarker discovery}
\usage{
LsOPLS(X, Y, No, Eta, Method)
}
\arguments{
\item{X}{A numeric matrix containing the spectra.}

\item{Y}{A numeric vector containing the target to predict.}

\item{No}{The number of orthogonal components to extract (may be optimalized previously).}

\item{Eta}{Parameter for the level of sparsity into the sPLS method}

\item{Method}{Parameter for the choice of the PLS algorithm ('pls2' or 'simpls')}
}
\value{
A print of the sPLS final step containing the list of final selected biomarkers.
}
\description{
Implementation of the Light-sparse-OPLS algorithm
Application of the sPLS method to a deflated matrix obtained by OPLS
(matrix without the Y-orthogonal components)
}
\examples{
data('iris')
X <- as.matrix(iris[,1:4])
Y <- as.numeric(iris[,5])
LsOPLS(X, Y, 2, 0.9, 'simpls')

}
