% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OPLDA.R
\name{OPLSDA}
\alias{OPLSDA}
\title{OPLS-DA for a two-class problem}
\usage{
OPLSDA(x, y, impT = FALSE, impG = FALSE, no = 2, nb = 15,
  out.path = ".")
}
\arguments{
\item{x}{A data matrix on which will be based the analysis.}

\item{y}{A numerical vector representing the class of individuals.}

\item{impT}{If \code{TRUE}, prints the results.}

\item{impG}{If \code{TRUE}, save results' graphics in pdf format.}

\item{no}{Number of orthogonal components to keep.}

\item{nb}{Number of biomarkers to select.}

\item{out.path}{Path to output the results' graphics}
}
\value{
A list with the following elements:
\describe{
  \item{\code{b.coef}}{Model coefficients}
  \item{\code{Tp}}{Predictive scores}
  \item{\code{Tortho}}{Orthogonal scores}
  \item{\code{Pp}}{Predictive loadings}
  \item{\code{Portho}}{Orthogonal loadings}
  \item{\code{Wortho}}{Orthogonal weights matrix}
  \item{\code{Selected.biomarkers}}{Vector of identified biomarkers}
  \item{\code{colmeans}}{Values of x column means}
}
}
\description{
OPLS-DA for a two-class problem. Only one response is allowed and the number of predictive components is fixed to 1 accordingly.
}
\examples{
data("DataSimul")
x = DataSimul[["x"]]
y = DataSimul[["y"]]
oplsda.res = OPLSDA(x=x, y=y, impT = FALSE,impG = FALSE, no=2, nb = 15, out.path = ".")

}

