% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PLSDA.R
\name{PLSDA}
\alias{PLSDA}
\title{PLS-DA}
\usage{
PLSDA(
  x,
  y,
  nLV = NULL,
  drawRMSEP = TRUE,
  validation = c("CV", "none", "LOO"),
  ...
)
}
\arguments{
\item{x}{A numeric matrix of spectral intensities.}

\item{y}{The real groups' memberships of the samples.}

\item{nLV}{The number of latent variables for PLS-DA.}

\item{drawRMSEP}{Logical, if \code{'TRUE'}, will draw plot of RMSEP values to decide on the number of latent variables.}

\item{validation}{The type of validation (see `pls::plsr()` for more info).}

\item{...}{additional arguments to be passed to `pls::plsr()` (eg: `segments`).}
}
\value{
A list with the following elements:
\describe{
  \item{\code{original.dataset}}{Original matrix of spectral intensities}
  \item{\code{Y}}{Dummy matrix representing the classes}
  \item{\code{nLV}}{Number of latent variables kept. If \code{NULL}, will be set to 3 in the finla model.}
  \item{\code{RMSEP}}{Root mean squared error of prediction }
  \item{\code{R2}}{Coefficient of multiple determination}
  \item{\code{Q2}}{Cross-validated coefficient of multiple determination}
  \item{\code{Q2cum}}{Cumulative cross-validated coefficient of multiple determination (over all response variables)}
  \item{\code{ExpVarY}}{Proportion of Y variance explained by the model}
  \item{\code{ExpVarX}}{Proportion of X variance explained by the model}
  \item{\code{coefficients}}{PLSDA coefficients}
  \item{\code{scores}}{PLSDA scores}
  \item{\code{loadings}}{PLSDA loadings}
  \item{\code{VIP}}{The variable VIPs}
   }
}
\description{
PLS-DA model for multiple classes (2 or more).
}
\examples{
data('HumanSerum')

PLSDA = PLSDA(x = HumanSerumSpectra,  y = ClassHS, nLV = NULL, drawRMSEP = TRUE, validation = "LOO")

PLSDA = PLSDA(x = HumanSerumSpectra,  y = ClassHS, nLV = NULL, drawRMSEP = TRUE, validation = "CV", segments = 5)

}
\author{
Manon Martin
}
