% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SVDforPCA.R
\name{SVDforPCA}
\alias{SVDforPCA}
\title{Singular value decomposition for PCA analysis}
\usage{
SVDforPCA(x, ncomp = min(dim(x)), roundval = TRUE)
}
\arguments{
\item{x}{A data matrix on which will be based the analysis.}

\item{ncomp}{Number of Principal Components.}

\item{roundval}{If \code{TRUE}, numeric variable names will be rounded to the 3rd decimal.}
}
\value{
A list with the following elements:
\describe{
  \item{\code{pcs}}{Scores}
  \item{\code{pcu}}{Normalized scores}
  \item{\code{pcv}}{Loadings}
  \item{\code{pcd}}{Singular values}
  \item{\code{var}}{Explained variance}
  \item{\code{cumvar}}{Cumulated explained variance}
  \item{\code{eigval}}{Eigenvalues}
}
}
\description{
PCA over a X matrix by singular value decomposition, the preprocessing involves the mean-centering of X.
}
\examples{
data("HumanSerum")
PCA.res = SVDforPCA(HumanSerumSpectra)
}

