% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SVDforPCA.R
\name{SVDforPCA}
\alias{SVDforPCA}
\title{Singular value decomposition for PCA analysis}
\usage{
SVDforPCA(x, ncomp = min(dim(x)))
}
\arguments{
\item{x}{A data matrix on which will be based the analysis.}

\item{ncomp}{Number of Principal Components.}
}
\value{
A list with the following elements:
\describe{
  \item{\code{scores}}{Scores}
  \item{\code{loadings}}{Loadings}
  \item{\code{eigval}}{Eigenvalues}
  \item{\code{pcd}}{Singular values}
  \item{\code{pcu}}{Normalized scores}
  \item{\code{var}}{Explained variance}
  \item{\code{cumvar}}{Cumulated explained variance}
 \item{\code{original.dataset}}{Original dataset}

}
}
\description{
PCA over a X matrix by singular value decomposition, the preprocessing involves the mean-centering of X.
}
\examples{
data('HumanSerum')
PCA.res = SVDforPCA(HumanSerumSpectra)
}
